pyproject.toml
[project]
name = "cognitive-qna-screening"
version = "0.1.0"
description = "Research prototype for cognitive Q&A screening (non-diagnostic)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.0",
  "openai>=1.30.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools]
packages = ["backend", "backend.app", "backend.app.instruments"]
